be/auth-pkce-demo/requirements.txt
# PKCE Demo Requirements

# FastAPI
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.7.0
sqlalchemy==2.0.36

# Security - bcrypt directly (replacing deprecated passlib)
bcrypt==4.2.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.22

# Templates
jinja2==3.1.4

# Redis
redis==5.2.1

# Logging
loguru>=0.7.3

# Email validation
email-validator==2.2.0

# Testing
pytest==8.3.4
httpx==0.28.1
